+++ python/python/tests/tests_redis_modules/test_json.py
@@ -152,6 +152,44 @@ class TestJson:
 
     @pytest.mark.parametrize("cluster_mode", [True, False])
     @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_del(self, redis_client: TRedisClient):
+        key = get_random_string(5)
+
+        json_value = {"a": 1.0, "b": {"a": 1, "b": 2.5, "c": True}}
+        assert await json.set(redis_client, key, "$", OuterJson.dumps(json_value)) == OK
+
+        assert await json.delete(redis_client, key, "$..a") == 2
+        assert await json.get(redis_client, key, "$..a") == "[]"
+
+        result = await json.get(redis_client, key, "$")
+        assert isinstance(result, str)
+        assert OuterJson.loads(result) == [{"b": {"b": 2.5, "c": True}}]
+
+        assert await json.delete(redis_client, key, "$") == 1
+        assert await json.delete(redis_client, key) == 0
+        assert await json.get(redis_client, key, "$") == None
+
+    @pytest.mark.parametrize("cluster_mode", [True, False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_forget(self, redis_client: TRedisClient):
+        key = get_random_string(5)
+
+        json_value = {"a": 1.0, "b": {"a": 1, "b": 2.5, "c": True}}
+        assert await json.set(redis_client, key, "$", OuterJson.dumps(json_value)) == OK
+
+        assert await json.forget(redis_client, key, "$..a") == 2
+        assert await json.get(redis_client, key, "$..a") == "[]"
+
+        result = await json.get(redis_client, key, "$")
+        assert isinstance(result, str)
+        assert OuterJson.loads(result) == [{"b": {"b": 2.5, "c": True}}]
+
+        assert await json.forget(redis_client, key, "$") == 1
+        assert await json.forget(redis_client, key) == 0
+        assert await json.get(redis_client, key, "$") == None
+
+    @pytest.mark.parametrize("cluster_mode", [True, False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
     async def test_json_toggle(self, redis_client: TRedisClient):
         key = get_random_string(10)
         json_value = {"bool": True, "nested": {"bool": False, "nested": {"bool": 10}}}